+++ src/pilo/core/threading/efficient_thread_pool.cpp
@@ -165,7 +165,7 @@ namespace pilo
             void efficient_thread_pool::set_running_handler(pool_callback_func_type hdl)
             {
                 this->_on_running_handler = hdl;
-                for (::pilo::i32_t i = this->_task_executor_count; i < this->_workers.size(); i++) {
+                for (::pilo::i32_t i = this->_task_executor_count; i < (int) this->_workers.size(); i++) {
                     this->_workers[i]->set_running_handler(hdl);
                 }
             }
